+++ apps/gbridge/widget.js
@@ -1,5 +1,4 @@
 (() => {
-
   const state = {
     music: "stop",
 
@@ -12,7 +11,16 @@
     scrollPos: 0
   };
 
+  function settings() {
+    let settings = require('Storage').readJSON("gbridge.json", true) || {};
+    if (!("showIcon" in settings)) {
+      settings.showIcon = true;
+    }
+    return settings
+  }
+
   function gbSend(message) {
+    Bluetooth.println("");
     Bluetooth.println(JSON.stringify(message));
   }
 
@@ -68,7 +76,7 @@
         var p = MAXCHARS;
         while (p > MAXCHARS - 8 && !" \t-_".includes(l[p]))
           p--;
-        if (p == MAXCHARS - 8) p = MAXCHARS;
+        if (p === MAXCHARS - 8) p = MAXCHARS;
         txt[i] = l.substr(0, p);
         txt.splice(i + 1, 0, l.substr(p));
       }
@@ -101,7 +109,7 @@
     const changed = state.music === event.state
     state.music = event.state
 
-    if (state.music == "play") {
+    if (state.music === "play") {
       showNotification(40, (y) => {
         g.setColor("#ffffff");
         g.drawImage(require("heatshrink").decompress(atob("jEYwILI/EAv/8gP/ARcMgOAASN8h+A/kfwP8n4CD/E/gHgjg/HA=")), 8, y + 8);
@@ -118,14 +126,14 @@
       }, changed);
     }
 
-    if (state.music == "pause") {
+    if (state.music === "pause") {
       hideNotification();
     }
   }
 
   function handleCallEvent(event) {
 
-    if (event.cmd == "accept") {
+    if (event.cmd === "accept") {
       showNotification(40, (y) => {
         g.setColor("#ffffff");
         g.drawImage(require("heatshrink").decompress(atob("jEYwIMJj4CCwACJh4CCCIMOAQMGAQMHAQMDAQMBCIMB4PwgHz/EAn4CBj4CBg4CBgACCAAw=")), 8, y + 8);
@@ -172,7 +180,7 @@
   });
 
   Bangle.on("swipe", (dir) => {
-    if (state.music == "play") {
+    if (state.music === "play") {
       const command = dir > 0 ? "next" : "previous"
       gbSend({ t: "music", n: command });
     }
@@ -191,10 +199,28 @@
     g.flip(); // turns screen on
   }
 
-  NRF.on("connect", changedConnectionState);
-  NRF.on("disconnect", changedConnectionState);
+  function reload() {
+    NRF.removeListener("connect", changedConnectionState);
+    NRF.removeListener("disconnect", changedConnectionState);
+    if (settings().showIcon) {
+      WIDGETS["gbridgew"].width = 24;
+      WIDGETS["gbridgew"].draw = draw;
+      NRF.on("connect", changedConnectionState);
+      NRF.on("disconnect", changedConnectionState);
+    } else {
+      WIDGETS["gbridgew"].width = 0;
+      WIDGETS["gbridgew"].draw = ()=>{};
+    }
+  }
 
-  WIDGETS["gbridgew"] = { area: "tl", width: 24, draw: draw };
+  WIDGETS["gbridgew"] = {area: "tl", width: 24, draw: draw, reload: reload};
+  reload();
 
-  gbSend({ t: "status", bat: E.getBattery() });
+  function sendBattery() {
+    gbSend({ t: "status", bat: E.getBattery() });
+  }
+
+  NRF.on("connect", () => setTimeout(sendBattery, 2000));
+  setInterval(sendBattery, 10*60*1000);
+  sendBattery();
 })();